voice/__main__.py
"""Allow running as: python -m voice dev"""

from voice.agent import entrypoint
from livekit.agents import WorkerOptions, cli

if __name__ == "__main__":
    cli.run_app(WorkerOptions(entrypoint_fnc=entrypoint))
